chore: modified the NewBlocks structure in the models.rs

+++ migrations/2024-11-07-100950_add_blocks/up.sql
@@ -1,6 +1,6 @@
 -- Your SQL goes here
 CREATE TABLE blocks (
     id SERIAL PRIMARY KEY,
-    block_number BIGINT NOT NULL,
+    block_number INT NOT NULL,
     timestamp TIMESTAMPTZ DEFAULT CURRENT_TIMESTAMP
 )
